pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["debian*", "binaries*", "build*", "dist*"]

[project]
name = "birdy"
version = "1.0.0"
dependencies = [
    "simplekml",
    "python-dateutil"
]
requires-python = ">=3.9"
authors = [
  {name = "Corey Forman (digitalsleuth)", email = "github@digitalsleuth.ca"}
]
maintainers = [
  {name = "Corey Forman (digitalsleuth)", email = "github@digitalsleuth.ca"}
]
description = "Python 3 script to parse the JSON ride data of the Bird Electric Scooter app"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["digital forensics", "dfir", "kml", "scooter", "bird", "birdy"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/digitalsleuth/birdy"

[project.scripts]
birdy = "birdy.birdy:main"
